feat: Add 'User Controller'

+++ app/models/UserModel.php
@@ -3,7 +3,14 @@
 namespace ghosty\taskmgr\models;
 
 use ghosty\taskmgr\database\DBHandle;
+use ghosty\taskmgr\dto\user\CreateUserDTO;
+use ghosty\taskmgr\dto\DTO;
+use ghosty\taskmgr\dto\user\FindUserByIdDTO;
+use ghosty\taskmgr\exceptions\AccessingNonExistentRecordException;
+use ghosty\taskmgr\exceptions\DatabaseException;
+use ghosty\taskmgr\logger\Severity;
 use ghosty\taskmgr\util\PasswordEncoder;
+use PDOException;
 
 class UserModel extends Model
 {
@@ -12,21 +19,29 @@ class UserModel extends Model
         parent::__construct($handle);
     }
 
-    public function insert(array $data): void
+    public function insert(CreateUserDTO | DTO $data): void
     {
         $template = "INSERT INTO users (username, password_hash, is_admin) VALUES (:username, :password_hash, :is_admin)";
 
         // encode user password to prevent storing plain_text
-        $data["password_hash"] = PasswordEncoder::encode($data["password_hash"]);
+        $password_hash = PasswordEncoder::encode($data->getPassword());
 
-        $this->handle->preparedStatement($template, $data);
+        try {
+            $this->handle->preparedStatement($template, $data->toArray() + ['password_hash' => $password_hash]);
+        } catch (PDOException $e) {
+            throw new DatabaseException($e->getMessage(), 500, Severity::WARNING, $e, __LINE__);
+        }
     }
 
-    public function findById(int $id): ?array
+    public function findById(DTO $data): ?array
     {
         $template = "SELECT * FROM users WHERE id = :id";
 
-        $result = $this->handle->preparedStatement($template, ['id' => $id])->fetch();
+        try {
+            $result = $this->handle->preparedStatement($template, $data->toArray())->fetch();
+        } catch (PDOException $e) {
+            throw new DatabaseException($e->getMessage(), 500, Severity::WARNING, $e, __LINE__);
+        }
 
         if (!$result) {
             return null;
@@ -35,33 +50,67 @@ class UserModel extends Model
 
     public function findAll(): array
     {
-        return $this->handle->query("SELECT * FROM users")->fetchAll();
+        try {
+            return $this->handle->query("SELECT * FROM users")->fetchAll();
+        } catch (PDOException $e) {
+            throw new DatabaseException($e->getMessage(), 500, Severity::WARNING, $e, __LINE__);
+        }
     }
 
-    public function update(float $id, array $data): void
+    public function update(DTO $data): void
     {
-        $this->handle->preparedStatement(
-            "UPDATE users SET 
-                 username = COALESCE(:username, username), 
-                 password_hash = COALESCE(:password_hash, password_hash) 
-                 WHERE id = :id",
-            $data + ['id' => $id]
-        );
+        if (!$this->existsById($data->getId())) {
+            throw new AccessingNonExistentRecordException($data->getId(), 'users', line: __LINE__);
+        }
+
+        $data = $data->toArray();
+
+        if (array_key_exists('new_password',$data)) {
+            $data += ['password_hash' => PasswordEncoder::encode($data['new_password'])];
+        }
+
+        try {
+            $this->handle->preparedStatement(
+                "UPDATE users SET 
+                 username = COALESCE(:new_username, username), 
+                 password_hash = COALESCE(:password_hash, password_hash), 
+                 is_admin = (:is_admin, is_admin)",
+                $data
+            );
+        } catch (PDOException $e) {
+            throw new DatabaseException($e->getMessage(), 500, Severity::WARNING, $e, __LINE__);
+        }
     }
 
-    public function delete(int $id): void
+    public function delete(FindUserByIdDTO | DTO $data): void
     {
+        if (!$this->existsById($data->getId())) {
+            throw new AccessingNonExistentRecordException($data->getId(), 'users', line: __LINE__);
+        }
+
         $this->handle->preparedStatement(
             "DELETE FROM users WHERE id = :id",
-            ['id' => $id]
+            ['id' => $data->getId()]
         );
     }
 
-    public function search(array $data): array
+    public function search(DTO $data): array
     {
         return $this->handle->preparedStatement(
             "SELECT * FROM users WHERE (:username IS NULL OR username LIKE :username)",
             $data
         )->fetchAll();
+    }
+
+    public function existsById(int $id): bool
+    {
+        try {
+            return $this->handle->preparedStatement(
+                "SELECT EXISTS(SELECT 1 FROM users WHERE id = :id)",
+                ['id' => $id]
+            )->fetchColumn();
+        } catch (PDOException $e) {
+            throw new DatabaseException($e->getMessage(), 500, Severity::WARNING, $e, __LINE__);
+        }
     }
 }
